+++ types/core/PubSub.d.ts
@@ -23,6 +23,7 @@ export class PubSub<T extends Record<string, unknown>> {
 }
 export namespace PubSub {
     let globalStore: Map<string | Symbol, PubSub<any>>;
+    let pendingDeps: Map<string | Symbol, Array<Function>>;
     let devMode: boolean;
 }
 export default PubSub;